Made 2 new functions: *ApplyNormalStress_SurfQuadratureStokes_FullFace() *AverageVolumeQuadraturePointsToSurfaceQuadraturePointsStokes()
The first one not only applies the lithostatic pressure to the traction vector but also computes and apply a normal shear stress. The normal shear stress is computed as follow:
*Use the velocity computed at previous time step to compute a strain rate
*Compute the normal strain rate to get a scalar as n[i]*E[i][j]*n[j]
*Compute the normal stress using the normal strain rate and the viscosity at surface quadrature points
This function requires the Stokes solution vector (X_stokes in the function) to access Uloc (and ultimately LA_Uloc to get velocity in the element). Therefore I had to change the definition of the function ModelApplyTractionFromLithoPressure() which is called in the driver to add the Stokes solution vector.

The second function makes a very simple assignement of viscosity and density values on Stokes surface quadrature points.
1- It gets the volume quadrature points and their attached values
2- For each face, make an average of the density and viscosity using a simple arithmetic mean
3- Push it on the surface quadrature points
This function is called in rheology.c in the function pTatin_EvaluateRheologyNonlinearities() so after the variables have been projected on the volume quadrature points
However, I still have an hesitation on what should be the index used to access the volume quadrature points. My guess is that it should be the global element number accessed via element_list[c] and not the local face number c, but I have a doubt. Need a double check.