Added 2 functions + 1 that is unused but I keep it for the record: 1- ModelApplyTractionFromLithoPressureStokesVolumeQuadrature() 2- ComputeLambdaFromStep() 3- ModelApplyTractionFromStokesVolumeQuadrature()
Function (1) Applies a Traction on the Stokes volume quadrature points to be used for the Neumann BCs. That function brings 2 new features:
1- It uses the viscosity from the volume quadrature points to compute the deviatoric stress tensor (itself on the volume quadrature points). This stress tensor is then integrated over the volume and divided by the volume of the element.
2- It uses a damping term lambda to progressively apply the normal deviatoric stress to the face (in addition of the lithostatic pressure which is always applied). This damping term stabilizes the BC avoiding for strong variations between two time steps that eventually lead to numerical instabilities.

Function (2) computes the damping term lambda based on ptatin steps:
Before the starting step, lambda value is lambda_min. At the ending step, lambda value is lambda_max. Between the two, lambda is interpolated from min to max depending on the time step.
The values of start_step, end_step, lambda_min and lambda_max are for now hard coded but I want to finally let the user access it through an option. But this requires to add it in drivers so before discussing it with Dave I let it that way.

Function (3) has the same purpose than function (1), but instead of using the lithostatic pressure it uses the actual pressure computed by the code. However, this implementation leads to very bad results for the velocity field. I do not know exactely why but:
It could come from the too bad initialization of the pressure field by the user in the ModelApplyInitialSolution() function (user model function). Anyway I tested several approaches with this function and none was successful so I decided to go back to the lithostatic pressure implementation.
I keep this function for the record but also for potential future use if an idea comes to solve the issues.