The pid and wil variables of the unassigned phase markers were erased because the marker_p was not called in this loop over material points but in the previous one. Therefore, the marker_p value was always the same and so the pid and wil variables were copied from the last marker_p accessed in the previous loop. That problem then lead to the cleaning of almost all of the newly inserted markers because the wil value is used to count how much markers lives in a cell and if the number of markers in a cell exceed the number of markers authorized they are removed by the function MPPC_SimpleRemoval(). Thus, since all the new markers have the same wil they are considered as being in the same cell (while they are not).
To correct that I use all_points[p].wil and all_points[p].pid to save the original wil and pid values in the exact same way than for the global and local coords. Small test on my machine was successful to keep the newly injected markers after passing through the cleaning from the population control.