*Seamount: Added a function to place a seamount (or a bathymetric anomaly) on the oceanic plate. This seamount is defined by a 2D Gaussian, all the parameters can be controlled with options. The surface is remeshed according to that seamount and it is compensated isostatically in depth.
*Organization:
Reorganized the ModelInitialize_SubductionOblique() function. It is splitted in 3 functions for easier use and readability:
1- ModelUserOptions(ModelSubductionObliqueCtx *data) => Model specific options as geometry, initial velocities, booleans etc...
2- ModelSetMaterialConstants(RheologyConstants *rheology, DataBucket materialconstants) => Assign material constants type and values from option file. Removed the useless array allocation that stored data for each region in the model.
3- ModelScaleQuantities(ModelSubductionObliqueCtx *data, RheologyConstants *rheology, DataBucket materialconstants) => Scale all the values that were given in the two functions above. Normally it should be the only place in the model where things are scaled. All other function calling model_ctx->variable should normally get scaled values.

Removed the sticky air related functionalities as it has never really worked.
Removed the distinction between ObliqueIC and ObliqueBC for the initial material geometry of material points as the only difference was the angle of the trench, so it can be controlled with one option.
Added an array to store the densities of each region in the data structure of the model. This array is used in the functions computing the isostatic topography. It is dynamically allocated in the ModelInitialize_SubductionOblique() and freed in ModelDestroy_SubductionOblique().