New data structure: GravityModel This data structure contains the following members: GravityType gravity_type; Vec         gravity_vec; PetscReal   gravity_const[3]; PetscReal   gravity_mag;
GravityType is an enum type currently containing the 'type' of gravity model we may want to use.
Currently only CONSTANT and RADIAL_CONSTANT are fully supported but types RADIAL_VAR, ARBITRARY and POISSON are already in the enum type (but can be removed)

The file gravity.c contains functions to:
- Create, allocate and zero the memory for the data structure GravityModel
- Free the memory of the data structure GravityModel
- Variety of getter and setter functions
- Set values of the gravity vector on quadrature points aswell as the body forces (density * gravity_vector).
Among the functions provided, 3 are of particular interest:
- pTatinCreateGravityModel(pTatinCtx ptatin, GravityType gtype, PetscReal scaling_factor, void *data)
For now, I see this function as the user's API to choose the gravity type and pass its scaling factor. The void *data parameter is here to deal with the fact that depending on the gravity type the necessary input from the user may differ (this can be discussed of course)
- pTatin_ApplyInitialStokesBodyForcesModel(pTatinCtx ptatin)
This function sets the gravity vector and body forces on the quadrature points
- pTatin_UpdateStokesBodyForcesModel(pTatinCtx ptatin)
This function updates the gravity vector and body forces on the quadrature points. For now they are both doing the same thing, we can discuss its usefulness.

Moreover, there is an if statement in the functions setting the values on the quadrature points, this if statement is here to indicate if we want to set only the gravity or both gravity and body forces. I did it that way to avoid having to loop twice over all the elements and quadrature points, so we can do both at the same time. That can also be discussed as I am not sure if it is the best solution.