Moved the GravityRadialConstant data structure in the c file. Note that any call to GravityRadialConstant outside of that file results in an incomplete type error. Thus all functions using this data structure must be defined inside this file.
Removed the (*set) function pointer to an implementation specific setter function. Easier for users.
Added a getter function attached to the function pointer (*get_gvec) to get the pointwise value of the gravity vector.
Note that it uses the PetscReal gvec_ptwise[3] located in the Gravity data structure as a buffer to ease memory management to avoid having floating malloc() and free() each time we need to get the pointwise vector.