Rename the GravityType enum members to add prefix GRAVITY_ to all of them. Rename the GravityModel data structure type to Gravity. Add PetscReal gvec_ptwise[3] to the data structure. This ease memory management when calling the (*get_gvec) functions. And because all types will have a get_gvec() function they will all require it so I put it in the parent Gravity structure instead of being in each one of the type specific data structure. Add  PetscErrorCode (*get_gvec)(Gravity,PetscInt,PetscReal*,PetscReal*,PetscReal**) function pointer. Function designed to get the pointwise gravity vector. Arguments are: 1 - Gravity. The Gravity data structure. 2 - PetscInt. The element (or cell) index. 3 - PetscReal*. The array containing the global coordinates of the point. 4 - PetscReal*. The array containing the local coordinates of the point. 5 - PetscReal**. The gravity vector array evaluated at the point.
Add const char *GravityTypeNames[] to store the textual (string) names of the different gravity types.
Add PetscErrorCode (*create_types[])(Gravity), an array to store the functions creating the data structure and setting the function pointers for each gravity type.
Removed the function that was calling the Create() functions of each type based on a switch and replaced it in GravitySetType() function.
Modification of the GravitySetType() function to call the proper Create() function based on the GravityType type passed by the user.
It makes use of the const char *GravityTypeNames[] to compare with the type passed and when matched it calls the Create() function stored in the array (*create_types[]).
This avoids a big switch statement and only requires to add an entry in *GravityTypeNames[] and (*create_types[]) when a new type is added.
Removed the (*set) function pointer for type specific functions. Easier to use as the users just have to call the setter according to the type they set.