Add ALE functionnality to Gene model
Create an enum type GeneMeshType to hold the type of mesh used. For now only MESH_EULERIAN and MESH_ALE are available.
The type defaults to MESH_EULERIAN but can be changed using the option -model_GENE3D_mesh_type {0,1}, 0: Eulerian, 1: ALE.
When using ALE mesh, the gmsh functionnality to tag surfaces for boundary conditions is no longer supported. Instead the user can only select full faces and their tag is already defined in the code by using the enum type HexElementFace such that:
- 0: east  = xmax = imax = HEX_FACE_Pxi
- 1: west  = xmin = imin = HEX_FACE_Nxi
- 2: north = ymax = jmax = HEX_FACE_Peta
- 3: south = ymin = jmin = HEX_FACE_Neta
- 4: front = zmax = kmax = HEX_FACE_Pzeta
- 5: back  = zmin = kmin = HEX_FACE_Nzeta
If a number greater than 5 or less than 0 is provided, it will error.
However, NO ERROR is given if one or more faces are not called by the user. This is mainly due to free-surface constraint being the natural BC with FEM but it means that the user should be careful when providing its BCs.

ALE mesh update uses velocity components to move the mesh but ALWAYS ensure to keep a rectangular domain, meaning that if a face has both inflow and outflow, the maximum outflow velocity controls the advection of the mesh.
By default all 3 components of the velocity are utilized to move the mesh but the user can decide to arbitrarily remove a component using the option(s) -model_GENE3D_mesh_ale_rm_component_{0,1,2} which will prevent any motion of the mesh in the given direction(s). Multiple directions can be given by calling the option for the desired component e.g., to remove y and z components: -model_GENE3D_mesh_ale_rm_component_1 -model_GENE3D_mesh_ale_rm_component_2

Finally, the functionality relying on expression evaluation has been updated to be able to treat the coordinates of the domain boundaries Ox,Lx,Oy,Ly,Oz,Lz as variables. This ensures that for a given velocity function, the value at boundaries stays constant when the boundaries move and that the gradient of the velocity function varies with the size of the domain (necessary for Navier-slip BCs).