Solve the Poisson pressure at each BC call
The solver uses an initial guess based on the previous solution.
This way, if the density has not changed no solve is done (the residual is already small enough) but if density has changed due to Boussinesq approximation and Stokes pressure changes during non-linear iterations, the residual on the poisson pressure is no more small and a new solve is performed.
It ensures that the pression applied by the Neumann BC is in agreement with the Stokes solution (at least the linear one).