Implement multiple heat sources in GENE + bugfix
Implementation of the functionality to be able to set different types of heat sources from options in model GENE.
Now source on material points and Decay types are also available.
The default behaviour is no more constant heat source per region but None i.e., no source.
Use options to activate the different sources.

Fix bug when applying source on material points:
When evaluating the heat source on material points in the EnergyFVEvaluateCoefficients_MaterialPoints() and EnergyEvaluateCoefficients_MaterialPoints() a sum is performed for each type of source activated on the given region.
However, when the type ENERGYSOURCE_USE_MATERIALPOINT_VALUE is selected in combination with other types, the previous implementation was summing every source type and set it on the material point.
Thus for a region with heat source from material points + constant + shear heating we have at time step n:

Hmp_{n} = Hc + Hmp_{n-1} + Hs

But at time step n+1:
H_{n+1} = Hc + Hmp_{n} + Hs

and so on, resulting in a sum of the previously summed heat sources at each time step and the source grows.
To fix this I added a heat_source_init variable in MPntPEnergy which is the variable that holds the initial value defined on markers by the user.