Quick fix for density tables
WARNING: this is a simple quick fix that will need to be better addressed in future.
This fix tries to ensure that if the density drops by a factor > 10 we conserve the previous density.
This factor is arbitrary and, even if it is unlikely that the density drops by a factor 10 from a physical process, we should make a more robust fix.