BETA: Add healing of plastic_strain
WARNING: work in progress, tests will be performed to decide if the functionality will be kept and how we should implement it.
For now its first version works as following:

In the plastic material parameters of the regions a healing_rate parameter is introduced. This parameter has the dimension of 1/time.
In the functions private_EvaluateRheologyNonlinearitiesMarkers_VPSTD (and its FV equivalent) on each marker of a region on which the plastic behaviour is activated (it can deform viscously but the region must allow plastic behaviour) the plastic strain is evaluated as

plastic_strain = plastic_strain - dt * healing_rate

with plastic_strain the value on the marker, dt the time step and healing_rate the healing_rate provided by the user for the region of the marker.

NOTE: this evaluation is performed at the END of the viscosity evaluation and will be thus taken into account for the next step. The value is then stored on the marker.

WARNING: With this implementation the plastic_strain is no more strain but damage. We may need to create a new marker variable to keep track of both because with the current implementation the plastic_strain can decrease and faults may disapear.