Add SOFTENING_LINEAR_THERMAL
Add a softening case based on temperature instead of plastic strain.
It is still used for plastic parameters but uses the temperature.
Because it is also linear it uses the same parameters than SOFTENING_LINEAR therefore I did not create a new material type structure but I re-use the one from the linear softening.