Add a user interface to select fields to output
Users can now choose which fields they want to output using options.
For now, the available options are:
-model_GENE3D_output_markercellp0_region
-model_GENE3D_output_markercellp0_viscosity
-model_GENE3D_output_markercellp0_density
-model_GENE3D_output_markercellp0_plastic_strain
-model_GENE3D_output_markercellp0_yield_indicator
-model_GENE3D_output_markercellp0_diffusivity
-model_GENE3D_output_markercellp0_heat_source
-model_GENE3D_output_markercellp0_damage

To avoid that anterior options files generate errors by missing these options a default behaviour is defined, if none of the above options is provided the fields outputed are:
  - region
  - viscosity
  - density
  - plastic strain

NO CHANGES:
Energy variables diffusivity and heat source are still outputed on the energyfv mesh.
Velocity and pressure are still outputed on the Q2 mesh.