Code improvement
Add a function to setup all the material constants data structure and memzero the MaterialData data structure.
Modifies how variables are accessed in the functions by reducing (when possible) the number of data structure access to reduce the operations on the memory by declaring temporary variables.
Declare temporary data structures as pointers to avoid memory copying of the content each time a function is called.