Add new neumann bc type
SC_DEVIATORIC_TRACTION allows constraining only the deviatoric part of the stress boundary conditions.
This formulation splits the stress into its deviatoric part and its volumetric part (the pressure).
The user provides a traction vector for the deviatoric stress and the pressure part is left unknown.
Two approaches are possible using this formulation:
  1- Provide the stress value directly
  2- Provide the strain-rate and use the quadrature points viscosity to compute the stress to impose
In both case the stress is the boundary condition, even in case 2 when providing the strain-rate, the stress is the value imposed, not the strain-rate.